+++ tianyi2_urdf/meshes/camera_head_link.STL
