+++ tianyi2_urdf/meshes/first_leg_pitch_link.STL
