+++ tianyi2_urdf/meshes/head_yaw_link.STL
