+++ tianyi2_urdf/meshes/left_index_1.STL
