+++ tianyi2_urdf/meshes/left_index_2.STL
