+++ tianyi2_urdf/meshes/left_little_1.STL
