+++ tianyi2_urdf/meshes/left_little_2.STL
