+++ tianyi2_urdf/meshes/left_middle_1.STL
