+++ tianyi2_urdf/meshes/left_middle_2.STL
