+++ tianyi2_urdf/meshes/left_ring_1.STL
