+++ tianyi2_urdf/meshes/left_ring_2.STL
