+++ tianyi2_urdf/meshes/left_tcp_link.STL
