+++ tianyi2_urdf/meshes/left_thumb_1.STL
