+++ tianyi2_urdf/meshes/left_thumb_2.STL
