+++ tianyi2_urdf/meshes/left_thumb_3.STL
