+++ tianyi2_urdf/meshes/left_thumb_4.STL
