+++ tianyi2_urdf/meshes/right_index_1.STL
