+++ tianyi2_urdf/meshes/right_index_2.STL
