+++ tianyi2_urdf/meshes/right_little_1.STL
