+++ tianyi2_urdf/meshes/right_little_2.STL
