+++ tianyi2_urdf/meshes/right_middle_1.STL
