+++ tianyi2_urdf/meshes/right_middle_2.STL
