+++ tianyi2_urdf/meshes/right_ring_1.STL
