+++ tianyi2_urdf/meshes/right_ring_2.STL
