+++ tianyi2_urdf/meshes/right_tcp_link.STL
