+++ tianyi2_urdf/meshes/right_thumb_1.STL
