+++ tianyi2_urdf/meshes/right_thumb_2.STL
