+++ tianyi2_urdf/meshes/right_thumb_3.STL
