+++ tianyi2_urdf/meshes/right_thumb_4.STL
