+++ tianyi2_urdf/meshes/second_leg_pitch_link.STL
