+++ tianyi2_urdf/meshes/shoulder_pitch_l_link.STL
