+++ tianyi2_urdf/meshes/shoulder_pitch_r_link.STL
