+++ tianyi2_urdf/meshes/shoulder_roll_l_link.STL
