+++ tianyi2_urdf/meshes/shoulder_roll_r_link.STL
