+++ tianyi2_urdf/meshes/wrist_pitch_l_link.STL
