+++ tianyi2_urdf/meshes/wrist_pitch_r_link.STL
