+++ tianyi2_urdf/script/check_joint_and_update.py
