+++ tianyi2_urdf/script/check_link_and_update.py
